requirements.txt
Flask
requests
Pillow
python-dotenv
certifi